feat: 식단 등록시 포인트 적립 기능 구현 및 기타 수정사항 반영

+++ backend/src/main/java/com/yunyun/meallog/pointshop/dao/OrderDao.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface OrderDao {
     int insertOrder(Order order);
     //사용자의 포인트 교환내역 조회
-    List<OrderHistoryDto> selectOrderHistoryByUserId(@Param("userId") int userId);
+    List<OrderHistoryDto> selectOrderHistoryByUserId(@Param("userId") Long userId);
 }
