+++ backend/src/main/java/com/yunyun/meallog/pointshop/dao/ProductDao.java
@@ -11,5 +11,5 @@ public interface ProductDao {
     // 특정 카테고리에 속하는 상품 목록 조회
     List<Product> selectProductsByCategory(@Param("category") String category);
     
-    Product selectProductById(int productId);
+    Product selectProductById(Long productId);
 }
